Updating the stats API
